@@ -52,8 +52,8 @@
     "po2json": "0.4.1",
     "punycode": "2.1.1",
     "querystring": "0.2.0",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
     "react-table": "7.3.2",
     "redux": "3.6.0",
     "shell-quote": "1.7.3",
