@@ -9,7 +9,7 @@
 import $ from 'jquery';
 import ko from 'knockout';
 import * as React from 'react';
-import * as ReactDOM from 'react-dom';
+import * as ReactDOMClient from 'react-dom/client';
 import mutate from 'mutate-cow';
 
 import Autocomplete from '../../common/components/Autocomplete';
@@ -230,7 +230,13 @@ class ArtistCreditEditor extends React.Component {
       .data('target', this.props.entity)
       .data('componentInst', this);
 
-    ReactDOM.render(
+    let bubbleRoot = $bubble.data('react-root');
+    if (!bubbleRoot) {
+      bubbleRoot = ReactDOMClient.createRoot($bubble[0]);
+      $bubble.data('react-root', bubbleRoot);
+    }
+
+    bubbleRoot.render(
       <ArtistCreditBubble
         addName={this.addName}
         artistCredit={this.state.artistCredit}
@@ -241,21 +247,22 @@ class ArtistCreditEditor extends React.Component {
         onNameChange={this.handleNameChange}
         pasteArtistCredit={this.pasteArtistCredit}
         removeName={this.removeName}
+        renderCallback={
+          show ? (() => {
+            this.positionBubble();
+
+            if (!bubbleWasVisible) {
+              $bubble.find(':input:eq(0)').focus();
+              $('#change-matching-artists').prop('checked', false);
+            }
+
+            if (callback) {
+              callback();
+            }
+          }) : null
+        }
         {...this.props}
       />,
-      $bubble[0],
-      show ? (() => {
-        this.positionBubble();
-
-        if (!bubbleWasVisible) {
-          $bubble.find(':input:eq(0)').focus();
-          $('#change-matching-artists').prop('checked', false);
-        }
-
-        if (callback) {
-          callback();
-        }
-      }) : null,
     );
   }
 
